Refactor (#15)
change Product icon: File to iconUrl: String

move impl file to correct package
