+++ src/main/kotlin/org/bibletranslationtools/fetcher/data/Product.kt
@@ -1,10 +1,8 @@
 package org.bibletranslationtools.fetcher.data
 
-import java.io.File
-
 data class Product(
     val slug: String,
     val titleKey: String,
     val descriptionKey: String,
-    val icon: File
+    val iconUrl: String
 )
