+++ src/main/kotlin/org/bibletranslationtools/fetcher/impl/repository/ProductCatalogImpl.kt
@@ -1,4 +1,4 @@
-package org.bibletranslationtools.fetcher.repository.implementations
+package org.bibletranslationtools.fetcher.impl.repository
 
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import com.fasterxml.jackson.module.kotlin.readValue
@@ -9,14 +9,6 @@ import org.bibletranslationtools.fetcher.repository.ProductCatalog
 
 class ProductCatalogImpl : ProductCatalog {
     private val productCatalogFileName = "product_catalog.json"
-
-    private data class ProductSchema(
-        val slug: String,
-        val titleKey: String,
-        val descriptionKey: String,
-        val iconUrl: String
-    )
-
     private val products: List<Product> = parseCatalog()
 
     override fun getAll(): List<Product> = this.products
@@ -29,15 +21,7 @@ class ProductCatalogImpl : ProductCatalog {
             return listOf()
         }
 
-        val productsFromSchema: List<ProductSchema> = jacksonObjectMapper().readValue(jsonFileTypes)
-        return productsFromSchema.map {
-            Product(
-                it.slug,
-                it.titleKey,
-                it.descriptionKey,
-                File(it.iconUrl)
-            )
-        }
+        return jacksonObjectMapper().readValue(jsonFileTypes)
     }
 
     @Throws(FileNotFoundException::class)
